move anilist list mutations to plugin
